Implement unbounce.
Initial implementation of unbounce and a unit test for the same.
When getBlob() is called and the response is satisfied from the
farStore, we first call putBlob() to move the blob to the
nearStore and then return the object to the caller. This has
an unfortunate property of blocking the client until we copy the blob
to the nearStore.