Pass BlobStore and container with StorageMetadata.
StorageMetadata on its own is not always sufficient (at least today)
to decide whether an object should be bounced or not. We may need to
interrogate the object store to make a decision. To resolve this, the
patch adds a ListBlobMetadata wrapper class that includes the
container name and the BlobStore instance. Both are required to look
up any additional metadata.

Refs: #29