Add a script to run the unit tests on EC2.
Commits a script that runs unit tests on EC2. The credentials are
retrieved from the BounceVerifier role and the bucket
bounce-test-config is expected to exist. The bucket is assumed to
contain the keys: github.key and test_creds.json.

The github.key contains the *private* key to access the Bounce git
repository.

The test_creds.json contains an array of credentials for providers to
use for tests. The tests support specifying only the "far" store and
the transient store is used for the "near" store.

Refs: #56