Implement reconcileObject() in BouncePolicy.
Bounce policies need to implement the reconciliation logic, as the
specifics of how to manipulate the objects are defined by the
policies. The BounceService in turn calls reconcileObject, as opposed
to using the FsckTask. This means that every time we click the
"bounce" button, reconciliation (if required) will also occur (in
addition to any bounce operations: write back, move, replication,
etc).