Add the migration policy.
The commit adds the migration policy and associated tests. The policy
implemnents migrations from "source" to "destination" blob stores (or
near to far as we configure them today).

In the process, to support reconciliation correctly without
assumptions about blob location, the BounceService run() method was
changed to perform the walk through the blob stores within a single
loop.