Implement the bounce button.
Implements the bounce button next to the virtual containers. The
button transitions to "Bouncing..." while the operation is in progress
and back to "Bounce" when the operation completes. Currently, it only
invokes the service. An additional improvement would be to allow for
ignoring the policy settings of when the bounce action should occur.