Show container status on the object store page.
We need to show the status for each container in the object store on
the object store page. The containers that we use are elided by the
server from the results. Containers that were "enhanced" (ones where a
cache, archive, or a migration target was added), are marked
accordingly and an "edit" button is displayed. Containers that were
unchanged are displayed with an "enhance" button.