Add status and vContainer ID to container listing.
When listing containers for a given provider, we need to return an
associated status and virtual container ID. Containers that are in use
(as a cache, archive, or migration target), are marked as INUSE. A
container for which a virtual container does not exist is marked as
UNCONFIGURED. A container for which a virtual container exists (and
which is setup with a cache, archive, or a migration target), is
marked as CONFIGURED.

The patch adds a small class to encapsulate the virtual container ID
and status tuple and constructs a map of containers names to the
tuples. The corresponding container object is constructed using these
values.