Implement statistics gathering with InfluxDB.
Adds a BounceStats class that is used to record statistics about the
operations happening on the system. The class supports only one kind
of series: "ops". The class exposes a logOperation() method, which
adds information about a particular op to be recorded in the stats
page. The class also exposes start()/stop() methods which are meant to
be used to start/stop the service that submits statistics to the local
influxdb instance. The statistics, if any are available, are submitted
every 30 seconds.