Create a NavBar controller.
The commit adds a NavBar controller, which is now used for displaying
the navigation bar. As part of this, active links are now displayed
correctly and the object store lists are updated when a new object
store is added.

This also fixes an issue where object stores were not displayed
correctly on every page, but only some pages.