Implement the bucket browser.
Implements the bucket browser. Containers are now links, clicking on
which opens a modal with the objects. For each object, we show its
locations, size, and name. The ContainerResource also exposes the
endpoint {object store ID}/container/{name} to query for the objects.
If the container is "enhanced", the listing contains the results from
querying the virtual container. To distinguish between the locations
an additional region is added -- FARTHER. This allows us to support
3-tier configurations.

Fixes #153