Log the backend ID, rather than the provider name.
Currently, we log the jclouds provider ID on PUT and GET operations.
This is not quite what we should be logging, as there could be
multiple jclouds providers with different regions or credentials. The
patch changes the logging to record the specific backend ID that is
logged (and adds a method to look up the provider by ID).