Track stats when performing bounce.
We need to track the space utilization and object count in each of the
object stores that bounce knows about. This is important for a number
of reasons:
1. Explaining the current cloud allocation
2. Understanding the progress of long-term bounce operations (e.g.
   migration)
3. Expressing any allocation-based storage policies (e.g. limiting the
   cache or primary storage size)
4. Demonstrating cost savings

To achieve this we add a new InfluxDB series to keep track of the
total number of objects and the total size of all objects in each of
the containers processed by the BounceService.