Use the container pool in tests.
We should use the container pool with any provider that is not
transient. The patch adds a method in the UtilsTest class to either
create a container or retrieve one from the pool. The method will only
create containers for the transient provider. All other providers
retrieve containers from the pool.

All tests are changed to call into this method rather than explicitly
create containers.

To clean up all outstanding containers, a RunListener is added to
invoke the ContainerPool.detroyAllContainers() method. The RunListener
is registered with surefire through maven.