Implement a container pool.
Implements a container pool that pre-allocates containers and
allocates additional containers out of band. The semantics are as
follows:
1. 10 containers are allocated initially
2. After 10 containers are allocated, the pool sleeps for 30 seconds
3. The pool maintains a thread pool and schedules a new container to
   be created after each container is checked out
4. Checked out containers must be removed by the caller (by calling
   delete container)

The pool does not work with the transient blob store, as it is using
the jclouds provider ID and endpoint to keep track of the blob store
associated with its container pool.