Implement a StoragePolicy.
Implements the StoragePolicy, where objects are expired if the total
size exceeds the object store capacity (provided by the user). We
should add high/low watermarks for the eviction policy and run the
bounce service possibly more frequently -- this will be done in
subsequent patches.

The policy operates on a per-day granularity and the objects are
evicted according to an LRU list.

Fixes #27