Add a script to create an OVA.
Adds a script that produces an OVA. The script uses VirtualBox
(specifically, VBoxManage utility). It also assumes the existence of
the bounce-base image. I will place the image into S3 and change the
script to pull it out of S3 when it runs (the image is a base Ubuntu
14.04.2 installation with docker, VirtualBox guest tools, and certain
configuration applied. The image is also pre-configured for SSH access
with the key that is being checked in.

The resulting ova is currently placed in /tmp. Will amend it to be
uploaded to S3 in the future.

Fixes #290